RSE: Refactor translation classes to a separate file
