+++ tests/test_rse_lfn2path.py
@@ -19,7 +19,7 @@ from configparser import NoOptionError, NoSectionError
 import pytest
 
 from rucio.common import config
-from rucio.rse.protocols.protocol import RSEDeterministicTranslation
+from rucio.rse.translation import RSEDeterministicTranslation
 
 
 @pytest.mark.noparallel(reason='uses pre-defined RSE, changes global configuration value')
